Tuck .create call inside the CreateProductWithCategoryService.
This narrows its interface and makes it easier to work with.

Note also the "renaming" of the doubles in the controller spec. This is
for compatibility with rspec-fire, which will be merged into rspec-mocks
3. By naming the double after the class it is representing, rspec can
validate that you are mocking methods that actually exist. (This isn't
enabled in this code base yet, but one day...)